fix: move helpers file into utils
