+++ packages/ui/src/breadcrumbs/navigation-dropdown.tsx
@@ -2,7 +2,7 @@
 
 import { CheckIcon } from "lucide-react";
 import * as React from "react";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 // ui
 import { CustomMenu, TContextMenuItem } from "../dropdowns";
 import { Tooltip } from "../tooltip";
