+++ packages/ui/src/breadcrumbs/navigation-search-dropdown.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useState } from "react";
 import { ICustomSearchSelectOption } from "@plane/types";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 import { CustomSearchSelect } from "../dropdowns";
 import { Tooltip } from "../tooltip";
 import { Breadcrumbs } from "./breadcrumbs";
