+++ packages/ui/src/card/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 import {
   ECardDirection,
   ECardSpacing,
