+++ packages/ui/src/dropdowns/custom-select.tsx
@@ -7,7 +7,7 @@ import { useOutsideClickDetector } from "@plane/hooks";
 // hooks
 import { useDropdownKeyDown } from "../hooks/use-dropdown-key-down";
 // helpers
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 // types
 import { ICustomSelectItemProps, ICustomSelectProps } from "./helper";
