+++ packages/ui/src/emoji/emoji-icon-picker-new.tsx
@@ -5,7 +5,7 @@ import EmojiPicker from "emoji-picker-react";
 // plane helpers
 import { useOutsideClickDetector } from "@plane/hooks";
 // helpers
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 // hooks
 import { LucideIconsList } from "./lucide-icons-list";
 // helpers
