+++ packages/ui/src/emoji/lucide-icons-list.tsx
@@ -2,7 +2,7 @@ import { Search } from "lucide-react";
 import React, { useEffect, useState } from "react";
 // local imports
 import { LUCIDE_ICONS_LIST } from "..";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 import { Input } from "../form-fields";
 import { InfoIcon } from "../icons";
 import { DEFAULT_COLORS, TIconsListProps, adjustColorForContrast } from "./emoji-icon-helper";
