+++ packages/ui/src/form-fields/checkbox.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 // helpers
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 
 export interface CheckboxProps extends React.InputHTMLAttributes<HTMLInputElement> {
   containerClassName?: string;
