+++ packages/ui/src/icons/priority-icon.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { AlertCircle, Ban, SignalHigh, SignalLow, SignalMedium } from "lucide-react";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 
 export type TIssuePriorities = "urgent" | "high" | "medium" | "low" | "none";
