+++ packages/ui/src/oauth/oauth-options.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 import { OAuthButton } from "./oauth-button";
 
 export type TOAuthOption = {
