+++ packages/ui/src/popovers/popover.tsx
@@ -2,7 +2,7 @@ import React, { Fragment, Ref, useState } from "react";
 import { usePopper } from "react-popper";
 import { Popover as HeadlessReactPopover, Transition } from "@headlessui/react";
 // helpers
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 // types
 import { TPopover } from "./types";
 import { EllipsisVertical } from "lucide-react";
