+++ packages/ui/src/scroll-area.tsx
@@ -1,7 +1,7 @@
 "use client";
 import * as RadixScrollArea from "@radix-ui/react-scroll-area";
 import React, { FC } from "react";
-import { cn } from "../helpers";
+import { cn } from "./utils";
 
 type TScrollAreaProps = {
   type?: "auto" | "always" | "scroll" | "hover";
