+++ packages/ui/src/tag/tag.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 import { ETagSize, ETagVariant, getTagStyle, TTagSize, TTagVariant } from "./helper";
 
 export interface TagProps extends React.ComponentProps<"div"> {
