+++ packages/ui/src/tooltip/tooltip.tsx
@@ -1,7 +1,7 @@
 import { Tooltip2 } from "@blueprintjs/popover2";
 import React, { useEffect, useRef, useState } from "react";
 // helpers
-import { cn } from "../../helpers";
+import { cn } from "../utils";
 
 export type TPosition =
   | "top"
