+++ packages/ui/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./classname";
 export * from "./icons";
